Free time getting/setting/passing works with time selector UI

+++ One Day at a Time Planner/Task.swift
@@ -40,5 +40,5 @@ class Task: CustomStringConvertible {
 
 struct todoInfo {
     var tasks: [Task]
-    var freeTime: [(Int, Int)]
+    var freeTime: [(Double, Double)]
 }
